Rework class creation dialog
It can now be used without the dialog, by calling the constructor with the new class name. This dialog class
takes care of generating the code, adding the rdfs:label, Prefered_Name, and FULL_SYN properties on the new
class. This change makes it suitable to use from the batch loader.