Remove ontology change listener and handle project open events
OWLEditorKitHooks are now loaded before GUIs, making it feasible to access the ClientSession object of
the protege client in the initialize call of NCIEditTab. This removes the need for the Ontology switch listener.
Also added a new OPEN_PROJECT event to the client, which is all that's need by EditTab to initialize properties.