New control logic for edittab
Now that tabs can be enabled/disabled in protege, workarounds have been removed and which
tabs are active can be controlled internally. For example once a user starts editing or retiring
a class, that class is pegged in the navigation pane and the user can't navigate away from it
until the operation is either completed or cancelled.