Add ability to unretire and unmerge during the pre-x phase
When a modeler has pre-merged or per-retired classes a new menu item appears on the right click of
the class in the pre-x branch. This can be used to undo the operation. Similar to how the retirements and merger
work, the user has to then save the class to commit the changes.